Refactor: GroupService로 이동
