+++ src/main/java/com/studypals/domain/groupManage/service/GroupService.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import com.studypals.domain.groupManage.dto.CreateGroupReq;
 import com.studypals.domain.groupManage.dto.GetGroupTagRes;
+import com.studypals.domain.groupManage.dto.GetGroupsRes;
 
 /**
  * GroupService 의 인터페이스입니다. 메서드를 정의합니다.
@@ -35,4 +36,11 @@ public interface GroupService {
      * @throws com.studypals.global.exceptions.exception.GroupException
      */
     Long createGroup(Long userId, CreateGroupReq dto);
+
+    /**
+     * 요청을 보낸 사용자가 속한 그룹의 정보를 조회합니다.
+     * @param userId 요청을 보낸 사용자
+     * @return 그룹 정보들
+     */
+    List<GetGroupsRes> getGroups(Long userId);
 }
